fix(web): align bot conversation targets
